Resize method in a separate action
